Adding user to documents

+++ backend/documents/documents/src/main/java/io/writeopia/api/documents/documents/repository/DocumentSqlBeDao.kt
@@ -37,8 +37,7 @@ class DocumentSqlBeDao(
                     icon = entity.icon?.let { MenuItem.Icon(it, entity.icon_tint) },
                     isLocked = entity.is_locked,
                 )
-            }
-            ?: emptyList()
+            } ?: emptyList()
 
     override suspend fun getLastUpdatedAt(userId: String): List<Document> =
         documentQueries?.selectLastUpdatedAt()
@@ -56,11 +55,15 @@ class DocumentSqlBeDao(
                     icon = entity.icon?.let { MenuItem.Icon(it, entity.icon_tint) },
                     isLocked = entity.is_locked
                 )
-            }
-            ?: emptyList()
+            } ?: emptyList()
 
     fun insertDocumentWithContent(document: Document) {
-        storyStepQueries?.deleteByDocumentId(document.id)
+        val result = documentQueries?.selectById(document.userId)?.executeAsOneOrNull()
+
+        if (result != null) {
+            storyStepQueries?.deleteByDocumentId(document.id)
+        }
+
         document.content.values.forEachIndexed { i, storyStep ->
             insertStoryStep(storyStep, i.toLong(), document.id)
         }
@@ -80,7 +83,8 @@ class DocumentSqlBeDao(
             parent_document_id = document.parentId,
             icon = document.icon?.label,
             icon_tint = document.icon?.tint,
-            is_locked = document.isLocked
+            is_locked = document.isLocked,
+            company_id = ""
         )
     }
 
@@ -104,12 +108,6 @@ class DocumentSqlBeDao(
                 spans = spans.joinToString(separator = ",") { it.toText() },
                 link_to_document = documentLink?.id
             )
-        }
-    }
-
-    fun insertDocuments(vararg documents: Document) {
-        documents.forEach { document ->
-            insertDocumentWithContent(document)
         }
     }
 
@@ -416,9 +414,10 @@ class DocumentSqlBeDao(
 
     fun loadDocumentsWithContentFolderIdAfterTime(
         folderId: String,
+        userId: String,
         time: Long
     ): List<Document> {
-        return documentQueries?.selectWithContentByFolderIdAfterTime(folderId, time)
+        return documentQueries?.selectWithContentByFolderIdAfterTime(folderId, time, userId)
             ?.executeAsList()
             ?.groupBy { it.id }
             ?.mapNotNull { (documentId, content) ->
